src, tests, c++ api: fix some warnings related to conversion to and from float

+++ examples/simple_training_net.c
@@ -78,7 +78,7 @@ static void init_net_data(float *data, uint32_t dim, const int *dims)
 {
     if (dim == 1) {
         for (int i = 0; i < dims[0]; ++i) {
-            data[i] = i % 1637;
+            data[i] = (float)(i % 1637);
         }
     } else if (dim == 4) {
         for (int in = 0; in < dims[0]; ++in) {
@@ -87,7 +87,7 @@ static void init_net_data(float *data, uint32_t dim, const int *dims)
                     for (int iw = 0; iw < dims[3]; ++iw) {
                         int indx = in * dims[1] * dims[2] * dims[3]
                                    + ic * dims[2] * dims[3] + ih * dims[3] + iw;
-                        data[indx] = indx % 1637;
+                        data[indx] = (float)(indx % 1637);
                     }
                 }
             }
